Refactor data source helpers
